Убрал использование UIKit у презентеров
